modify test to test more scenarios
